refactor: split settings page scripts into modules
