+++ tests/static/test_modal_accessibility_guards.py
@@ -53,7 +53,7 @@ def test_plugin_page_open_modal_focuses_first_focusable():
 
 def test_settings_page_reboot_shutdown_modals_handle_escape():
     """JTN-652: /settings reboot + shutdown confirm modals close on Escape."""
-    content = _read_script("settings_page.js")
+    content = _read_script("settings/modals.js")
 
     assert 'event.key !== "Escape"' in content
     assert "rebootConfirmModal" in content
@@ -63,10 +63,10 @@ def test_settings_page_reboot_shutdown_modals_handle_escape():
 def test_settings_page_device_action_modals_manage_focus():
     """JTN-652: /settings confirm modals move focus in on open and restore
     focus to the trigger on close."""
-    content = _read_script("settings_page.js")
+    content = _read_script("settings/modals.js")
 
     assert "focusable.focus()" in content
-    assert "_lastDeviceActionTrigger" in content
+    assert "DeviceActionTrigger" in content
 
 
 def test_response_modal_restores_focus_to_trigger():
