+++ tests/static/test_settings_interval_display.py
@@ -2,6 +2,8 @@
 
 from pathlib import Path
 
+JS_PATH = Path("src/static/scripts/settings/form.js")
+
 
 def test_populate_interval_fields_clamps_sub_minute():
     """JTN-245: Values < 60 seconds must not display as '0 minutes'.
@@ -9,13 +11,13 @@ def test_populate_interval_fields_clamps_sub_minute():
     Math.max(1, intervalInMinutes) ensures that sub-minute intervals
     are shown as 1 minute rather than 0 minutes (which fails min="1" validation).
     """
-    js = Path("src/static/scripts/settings_page.js").read_text()
+    js = JS_PATH.read_text()
     assert "Math.max(1, intervalInMinutes)" in js
 
 
 def test_populate_interval_fields_hours_branch_unchanged():
     """The hours branch should still use the raw hour value without clamping."""
-    js = Path("src/static/scripts/settings_page.js").read_text()
+    js = JS_PATH.read_text()
     assert "intervalInput.value = String(intervalInHours)" in js
     assert 'unitSelect.value = "hour"' in js
